[Feature #164054693] Inherit from single base test class
