+++ tests/v2/test_canditate_register.py
@@ -6,72 +6,50 @@ from app import app
 from config import app_config
 from app.api.v2.models.db import init_db
 
+from tests.v2.base_test import BaseTestClass
+
 import jwt
 import os
 KEY = os.getenv('SECRET_KEY')
 ADMIN_TOKEN = os.getenv("ADMIN_TOKEN")
 
 
-class BaseTestClass(unittest.TestCase):
-    """
-    Setting up tests
-    """
-
-    def setUp(self):
-        self.app = app("testing")
-        self.client = self.app.test_client()
-        self.app.config['TESTING'] = True
-        init_db()
-
-        self.admintoken = ADMIN_TOKEN
-
-        self.newoffice = {
-            "type": "Governor",
-            "name": "Governor Narok County"
-        }
-        self.newparty = {
-            "name": "Party1",
-            "hqAddress": "Nairobi",
-            "logoUrl": ""
-        }
-
-        self.new_user_as_politician = {
-            "username": "Tevyn",
-            "firstname": "Tevin",
-            "lastname": "Gach",
-            "email": "tevinku@gmail.com",
-            "phone": "0735464438",
-            "othername": "Thuku",
-            "password": "Tevin1995",
-            "retypedpassword": "Tevin1995",
-            "passportUrl": "http",
-            "isPolitician": True
-        }
-
-    def tearDown(self):
-        """Clear the db after tests finish running"""
-        self.app.testing = False
-        init_db()
-
-
 class TestUserEndpoints(BaseTestClass):
     def CreatePoliticianUser(self):
         return self.client.post("api/v2/auth/signup",
-                                data=json.dumps(self.new_user_as_politician),
+                                data=json.dumps({
+                                    "username": "Tevyn",
+                                    "firstname": "Tevin",
+                                    "lastname": "Gach",
+                                    "email": "tevinku@gmail.com",
+                                    "phone": "0735464438",
+                                    "othername": "Thuku",
+                                    "password": "Tevin1995",
+                                    "retypedpassword": "Tevin1995",
+                                    "passportUrl": "http",
+                                    "isPolitician": True
+                                }),
                                 content_type="application/json")
 
     def AdminCreateOffice(self):
         return self.client.post(
             "api/v2/offices",
-            data=json.dumps(self.newoffice),
-            headers={'x-access-token': self.admintoken},
+            data=json.dumps({
+                "type": "Governor",
+                "name": "Governor Narok County"
+            }),
+            headers={'x-access-token': ADMIN_TOKEN},
             content_type="application/json")
 
     def AdminCreateParty(self):
         return self.client.post(
             "api/v2/parties",
-            data=json.dumps(self.newparty),
-            headers={'x-access-token': self.admintoken},
+            data=json.dumps({
+                "name": "Party1",
+                "hqAddress": "Nairobi",
+                "logoUrl": ""
+            }),
+            headers={'x-access-token': ADMIN_TOKEN},
             content_type="application/json")
 
     def AdminRegisterCandidate(self):
@@ -80,7 +58,7 @@ class TestUserEndpoints(BaseTestClass):
                                     "office": 1,
                                     "user": 2
                                 }),
-                                headers={'x-access-token': self.admintoken},
+                                headers={'x-access-token': ADMIN_TOKEN},
                                 content_type="application/json")
 
     def test_admin_register_politician_candidate(self):
@@ -105,7 +83,7 @@ class TestUserEndpoints(BaseTestClass):
         res = self.client.post("api/v2/offices/1/register",
                                data=json.dumps({
                                }),
-                               headers={'x-access-token': self.admintoken},
+                               headers={'x-access-token': ADMIN_TOKEN},
                                content_type="application/json")
         self.assertEqual(res.status_code, 400)
 
@@ -117,7 +95,7 @@ class TestUserEndpoints(BaseTestClass):
                                data=json.dumps({
                                    "user": 2
                                }),
-                               headers={'x-access-token': self.admintoken},
+                               headers={'x-access-token': ADMIN_TOKEN},
                                content_type="application/json")
         self.assertEqual(res.status_code, 404)
 
@@ -129,7 +107,7 @@ class TestUserEndpoints(BaseTestClass):
                                data=json.dumps({
                                    "user": 20
                                }),
-                               headers={'x-access-token': self.admintoken},
+                               headers={'x-access-token': ADMIN_TOKEN},
                                content_type="application/json")
         self.assertEqual(res.status_code, 404)
 
@@ -156,6 +134,6 @@ class TestUserEndpoints(BaseTestClass):
                                    "office": 1,
                                    "user": ""
                                }),
-                               headers={'x-access-token': self.admintoken},
+                               headers={'x-access-token': ADMIN_TOKEN},
                                content_type="application/json")
         self.assertEqual(res.status_code, 400)
