+++ tests/v2/test_parties.py
@@ -6,71 +6,39 @@ from app import app
 from config import app_config
 from app.api.v2.models.db import init_db
 
+from .base_test import BaseTestClass
 import jwt
 import os
 KEY = os.getenv('SECRET_KEY')
 ADMIN_TOKEN = os.getenv("ADMIN_TOKEN")
 
 
-class BaseTestClass(unittest.TestCase):
-    """
-    Setting up tests
-    """
-
-    def setUp(self):
-        self.app = app("testing")
-        self.client = self.app.test_client()
-        self.app.config['TESTING'] = True
-        init_db()
-
-        self.new_user = {
-            "username": "Tevyn",
-            "firstname": "Tevin",
-            "lastname": "Gach",
-            "email": "tevinku@gmail.com",
-            "phone": "0735464438",
-            "othername": "Thuku",
-            "password": "Tevin1995",
-            "retypedpassword": "Tevin1995",
-            "passportUrl": "http",
-            "isPolitician": False
-        }
-
-        self.login_new_user = {
-            "email": "tevinku@gmail.com",
-            "password": "Tevin1995"
-        }
-
-        self.newparty = {
-            "name": "Party1",
-            "hqAddress": "Nairobi",
-            "logoUrl": ""
-        }
-
-        self.missinginfoparty = {
-            "hqAddress": "Nairobi",
-            "logoUrl": ""
-        }
-
-        self.admintoken = ADMIN_TOKEN
-
-    def tearDown(self):
-        """Clear the db after tests finish running"""
-        self.app.testing = False
-        init_db()
-
-
 class TestPartiesFunctionality(BaseTestClass):
     def CreateUser(self):
         return self.client.post("api/v2/auth/signup",
-                                data=json.dumps(self.new_user),
+                                data=json.dumps({
+                                    "username": "Tevyn",
+                                    "firstname": "Tevin",
+                                    "lastname": "Gach",
+                                    "email": "tevinku@gmail.com",
+                                    "phone": "0735464438",
+                                    "othername": "Thuku",
+                                    "password": "Tevin1995",
+                                    "retypedpassword": "Tevin1995",
+                                    "passportUrl": "http",
+                                    "isPolitician": False
+                                }),
                                 content_type="application/json")
 
     def AdminPostParty(self):
         return self.client.post(
             "api/v2/parties",
-            data=json.dumps(self.newparty),
-            headers={'x-access-token': self.admintoken},
+            data=json.dumps({
+                "name": "Party1",
+                "hqAddress": "Nairobi",
+                "logoUrl": ""
+            }),
+            headers={'x-access-token': ADMIN_TOKEN},
             content_type="application/json")
 
     def test_fetching_all_parties(self):
@@ -94,7 +62,7 @@ class TestPartiesFunctionality(BaseTestClass):
                             "hqAddress": "",
                             "logoUrl": ""
                             }),
-            headers={'x-access-token': self.admintoken},
+            headers={'x-access-token': ADMIN_TOKEN},
             content_type="application/json")
 
         self.assertEqual(res.status_code, 400)
@@ -102,7 +70,11 @@ class TestPartiesFunctionality(BaseTestClass):
     def test_posting_party_anonymously(self):
         response = self.client.post(
             "api/v2/parties",
-            data=json.dumps(self.newparty),
+            data=json.dumps({
+                "name": "Party1",
+                "hqAddress": "Nairobi",
+                "logoUrl": ""
+            }),
             content_type="application/json"
         )
         self.assertEqual(response.status_code, 401)
@@ -112,11 +84,18 @@ class TestPartiesFunctionality(BaseTestClass):
     def test_ordinary_user_attempting_to_create_a_party(self):
         self.CreateUser()
         userdata = self.client.post(
-            "api/v2/auth/signin", data=json.dumps(self.login_new_user), content_type="application/json")
+            "api/v2/auth/signin", data=json.dumps({
+                "email": "tevinku@gmail.com",
+                "password": "Tevin1995"
+            }), content_type="application/json")
         usertoken = json.loads(userdata.data.decode("utf-8"))["data"]["token"]
         response = self.client.post(
             "api/v2/parties",
-            data=json.dumps(self.newparty),
+            data=json.dumps({
+                "name": "Party1",
+                "hqAddress": "Nairobi",
+                "logoUrl": ""
+            }),
             headers={'x-access-token': usertoken},
             content_type="application/json")
         self.assertEqual(response.status_code, 401)
@@ -126,8 +105,11 @@ class TestPartiesFunctionality(BaseTestClass):
     def test_admin_post_party_with_missing_fields(self):
         response = self.client.post(
             "api/v2/parties",
-            data=json.dumps(self.missinginfoparty),
-            headers={'x-access-token': self.admintoken},
+            data=json.dumps({
+                "hqAddress": "Nairobi",
+                "logoUrl": ""
+            }),
+            headers={'x-access-token': ADMIN_TOKEN},
             content_type="application/json")
         self.assertEqual(response.status_code, 400)
         result = json.loads(response.data.decode("utf-8"))
@@ -141,7 +123,7 @@ class TestPartiesFunctionality(BaseTestClass):
                 "hqAddress": "Int name",
                 "logoUrl": "logoUrl"
             }),
-            headers={'x-access-token': self.admintoken},
+            headers={'x-access-token': ADMIN_TOKEN},
             content_type="application/json")
         self.assertEqual(response.status_code, 400)
 
@@ -150,7 +132,11 @@ class TestPartiesFunctionality(BaseTestClass):
             {"email": "ninja@gmail.com"}, KEY, algorithm='HS256')
         response = self.client.post(
             "api/v2/parties",
-            data=json.dumps(self.newparty),
+            data=json.dumps({
+                "name": "Party1",
+                "hqAddress": "Nairobi",
+                "logoUrl": ""
+            }),
             headers={'x-access-token': mocktoken},
             content_type="application/json")
         self.assertEqual(response.status_code, 401)
@@ -192,7 +178,7 @@ class TestPartiesFunctionality(BaseTestClass):
         self.AdminPostParty()
         res = self.client.patch("api/v2/parties/1/name",
                                 data=json.dumps({"name": "Party A"}),
-                                headers={'x-access-token': self.admintoken},
+                                headers={'x-access-token': ADMIN_TOKEN},
                                 content_type="application/json")
         self.assertEqual(res.status_code, 200)
         dataresponse = json.loads(res.data.decode("utf-8"))
@@ -205,7 +191,7 @@ class TestPartiesFunctionality(BaseTestClass):
         self.AdminPostParty()
         res = self.client.patch("api/v2/parties/1/name",
                                 data=json.dumps({"name": ""}),
-                                headers={'x-access-token': self.admintoken},
+                                headers={'x-access-token': ADMIN_TOKEN},
                                 content_type="application/json")
         self.assertEqual(res.status_code, 400)
 
@@ -213,7 +199,7 @@ class TestPartiesFunctionality(BaseTestClass):
         self.AdminPostParty()
         res = self.client.patch("api/v2/parties/1/name",
                                 data=json.dumps({"name": 12}),
-                                headers={'x-access-token': self.admintoken},
+                                headers={'x-access-token': ADMIN_TOKEN},
                                 content_type="application/json")
         self.assertEqual(res.status_code, 400)
 
@@ -221,7 +207,7 @@ class TestPartiesFunctionality(BaseTestClass):
         self.AdminPostParty()
         self.client.patch("api/v2/parties/1/name",
                           data=json.dumps({"name": "Party A"}),
-                          headers={'x-access-token': self.admintoken},
+                          headers={'x-access-token': ADMIN_TOKEN},
                           content_type="application/json")
         response = self.client.get("api/v2/parties",
                                    content_type="application/json")
@@ -237,7 +223,7 @@ class TestPartiesFunctionality(BaseTestClass):
     def test_admin_updating_non_existent_party(self):
         res = self.client.patch("api/v2/parties/20/name",
                                 data=json.dumps({"name": "Party 20 num"}),
-                                headers={'x-access-token': self.admintoken},
+                                headers={'x-access-token': ADMIN_TOKEN},
                                 content_type="application/json")
         self.assertEqual(res.status_code, 404)
         dataresponse = json.loads(res.data.decode("utf-8"))
@@ -248,7 +234,7 @@ class TestPartiesFunctionality(BaseTestClass):
         res = self.client.patch("api/v2/parties/1/name",
                                 data=json.dumps(
                                     {"newname": "Missing new name"}),
-                                headers={'x-access-token': self.admintoken},
+                                headers={'x-access-token': ADMIN_TOKEN},
                                 content_type="application/json")
         self.assertEqual(res.status_code, 400)
         dataresponse = json.loads(res.data.decode("utf-8"))
@@ -269,7 +255,7 @@ class TestPartiesFunctionality(BaseTestClass):
         res = self.client.delete(
             "/api/v2/parties/{}".format(1),
             content_type="application/json",
-            headers={'x-access-token': self.admintoken}
+            headers={'x-access-token': ADMIN_TOKEN}
         )
         self.assertEqual(res.status_code, 200)
 
@@ -278,7 +264,7 @@ class TestPartiesFunctionality(BaseTestClass):
         res = self.client.delete(
             "/api/v2/parties/{}".format(100),
             content_type="application/json",
-            headers={"x-access-token": self.admintoken}
+            headers={"x-access-token": ADMIN_TOKEN}
         )
         self.assertEqual(res.status_code, 404)
